fix editing component for linked files in dataset table
