+++ react/src/Datasets/components/DatasetTable.tsx
@@ -2,14 +2,13 @@ import React, { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { makeStyles, Chip, IconButton, TextField } from "@material-ui/core";
 import { PlayArrow, Delete, Cancel, Visibility } from "@material-ui/icons";
-import { Autocomplete } from "@material-ui/lab";
 import MaterialTable, { MTableToolbar } from "material-table";
 import { useSnackbar } from "notistack";
 import { toKeyValue, formatDateString, exportCSV, rowDiff } from "../../functions";
-import { KeyValue, Dataset, Pipeline } from "../../typings";
+import { KeyValue, Dataset, Pipeline, Option } from "../../typings";
 import AnalysisRunnerDialog from "./AnalysisRunnerDialog";
 import DatasetInfoDialog from "./DatasetInfoDialog";
-import { Note } from "../../components";
+import { Note, FileLinkingComponent } from "../../components";
 import LinkedFilesButton from "./LinkedFilesButton";
 import { useFetchCache } from "../../contexts/fetchCache";
 
@@ -165,15 +164,13 @@ export default function DatasetTable({ isAdmin }: DatasetTableProps) {
                         customSort: (a, b) => a.linked_files.length - b.linked_files.length,
                         render: data => <LinkedFilesButton fileNames={data.linked_files} />,
                         editComponent: props => (
-                            <Autocomplete
-                                selectOnFocus
-                                clearOnBlur
-                                handleHomeEndKeys
-                                autoHighlight
-                                onChange={(event, newValue) => props.onChange(newValue)}
-                                options={files}
-                                value={props.value}
-                                renderInput={params => <TextField {...params} variant="standard" />}
+                            <FileLinkingComponent
+                                values={props.rowData.linked_files}
+                                options={files.map(
+                                    file => ({ title: file, inputValue: file } as Option)
+                                )}
+                                onEdit={(newValue: string[]) => props.onChange(newValue)}
+                                disableTooltip
                             />
                         ),
                     },
@@ -258,6 +255,7 @@ export default function DatasetTable({ isAdmin }: DatasetTableProps) {
                             <div className={classes.chipBar}>
                                 {[...new Set(datasets.map(e => e.dataset_type))].map(type => (
                                     <Chip
+                                        key={type}
                                         label={type}
                                         onClick={() => setDatasetTypeFilter([type])}
                                         clickable
